+++ api/python/src/ELF/pyELF.cpp
@@ -26,6 +26,7 @@ void init_python_module(py::module& m) {
 }
 
 void init_objects(py::module& m) {
+  CREATE(ParserConfig, m);
   CREATE(Parser, m);
   CREATE(SymbolVersion, m);
   CREATE(Binary, m);